Introduce webui tab in stable diffusion
